* Fix format string issues. * Fix type of FCHOOKDLL_API_VERSION macro constant. * Fix AppSoundsRegistry issue. * NTray: Fix SAL annotation warning.

+++ FreeCompose/branches/new-approach/SharedInclude/FCHookDll.h
@@ -10,7 +10,7 @@ extern "C" {
 #define FCHOOKDLL_API __declspec(dllimport)
 #endif
 
-#define FCHOOKDLL_API_VERSION 0x0030
+#define FCHOOKDLL_API_VERSION ((DWORD) 0x0030)
 
 //==============================================================================
 // Types
